refactor: create account hook for accessing user info from session
